feat: don't need janky editor state setting anymore
