+++ apps/web/app/space/(entity)/[id]/[entityId]/layout.tsx
@@ -7,7 +7,7 @@ import { Metadata } from 'next';
 import { Entity, Relation } from '~/core/io/dto/entities';
 import { fetchBlocks } from '~/core/io/fetch-blocks';
 import { EntityId, TypeId } from '~/core/io/schema';
-import { EditorProvider } from '~/core/state/editor-store';
+import { EditorProvider } from '~/core/state/editor/editor-provider';
 import { EntityStoreProvider } from '~/core/state/entity-page-store/entity-store-provider';
 import { TypesStoreServerContainer } from '~/core/state/types-store/types-store-server-container';
 import { Entities } from '~/core/utils/entity';
